@@ -79,16 +84,11 @@
     "style-loader": "^0.23.1",
     "terser-webpack-plugin": "^1.2.3",
     "thread-loader": "^2.1.2",
-    "ts-loader": "^5.3.3",
+    "ts-loader": "^6.0.2",
     "ts-node": "^8.0.3",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "tslint": "^5.13.1",
-    "tslint-config-airbnb": "^5.11.1",
-    "tslint-config-prettier": "^1.13.0",
-    "tslint-eslint-rules": "^5.3.1",
-    "tslint-react": "^4.0.0",
-    "typescript": "^3.4.2",
-    "url-loader": "1.1.2",
+    "typescript": "^3.5.1",
+    "url-loader": "2.0.0",
     "webpack": "^4.29.6",
     "webpack-cli": "^3.0.8",
     "webpack-dev-server": "^3.2.1",
