@@ -5,15 +5,13 @@ import Modal from 'src/components/elements/Modal';
 interface Props {
   visible?: boolean;
 }
-class Register extends React.Component<Props> {
+class Register extends React.Component<Props, any> {
   state = {
     visible: this.props.visible
   };
 
   toggleEvent = () => {
-    this.setState({
-      visible: !this.state.visible
-    });
+    this.setState(prevState => ({ visible: !prevState.visible }));
   };
 
   render() {
