@@ -2,10 +2,6 @@ import * as React from 'react';
 
 import Error from './Error';
 
-class Page extends React.Component {
-  render() {
-    return <Error />;
-  }
-}
+const Page = () => <Error />;
 
 export default Page;
