@@ -6,7 +6,7 @@ import AsyncComponent from 'src/components/shared/AsyncComponent';
 import RouteStore from 'src/state/stores/routeStore';
 import CustomRoutes from './CustomRoutes';
 import RouteGenerator from './RouteGenerator';
-import { setPageRouteSuspenseAsync } from './routeHelpers';
+import setPageRouteSuspenseAsync from './routeHelpers';
 
 const Home = setPageRouteSuspenseAsync('/Home');
 
@@ -21,8 +21,10 @@ class Routes extends React.Component {
         <Route exact path="/">
           {Home}
         </Route>
+
         {RouteGenerator(this.RouteStore)}
         {CustomRoutes()}
+
         <Route>
           <AsyncComponent
             importStatement={import('src/components/shared/NotFound')}
