@@ -1,6 +1,6 @@
 import { observer } from 'mobx-react';
 import * as React from 'react';
-import { Card, Segment } from 'semantic-ui-react';
+import { Segment } from 'semantic-ui-react';
 import BirdStore from 'src/state/stores/games/birdStore';
 
 interface Props {
@@ -9,7 +9,7 @@ interface Props {
 
 const Debug = observer((props: Props) => (
   <Segment basic>
-    <h2>Debwug</h2>
+    <h2>Debug</h2>
     <code>{JSON.stringify(props.stats, null, 4)}</code>
   </Segment>
 ));
