@@ -1,5 +1,4 @@
-import { observer } from 'mobx-react';
-import React from 'react';
+import * as React from 'react';
 
 import { Group } from 'react-konva';
 import Pipe from './Pipe';
