@@ -11,24 +11,22 @@ import Card from 'src/components/elements/Card';
 
 interface Props {
   clearScore: () => void;
-  stats: Array<{
+  stats: {
     winner: string;
     gameNumber: number;
     totalMoves: number;
     boxOrder: [number];
-  }>;
+  }[];
 }
 
 interface State {
   active: string;
-  filteredList: object[];
 }
 
 class ScoreCard extends React.Component<Props, State> {
   // default active state
   state = {
-    active: '',
-    filteredList: this.props.stats
+    active: ''
   };
 
   filterList = (stats, content) => {
@@ -45,17 +43,13 @@ class ScoreCard extends React.Component<Props, State> {
   };
 
   activeChange = (e, { content }) => {
-    const { stats } = this.props;
-    const list = this.filterList(stats, content);
     if (content === this.state.active) {
       this.setState({
-        active: '',
-        filteredList: list
+        active: ''
       });
     } else {
       this.setState({
-        active: content,
-        filteredList: list
+        active: content
       });
     }
   };
