@@ -1,13 +1,9 @@
 import * as React from 'react';
 
-class SuspenseTest extends React.Component<any, any> {
-  render() {
-    return (
-      <div>
-        <p>www</p>
-      </div>
-    );
-  }
-}
+const SuspenseTest = () => (
+  <div>
+    <p>www</p>
+  </div>
+);
 
 export default SuspenseTest;
