@@ -6,13 +6,15 @@ const hostname = window && window.location && window.location.hostname;
 // determine if app is staging
 const staging = hostname === process.env.STAGING;
 
-let rootUrl = '';
+let internalRrootUrl = '';
 if (mode === 'development') {
-  rootUrl = 'http://localhost:5000/api';
+  internalRrootUrl = 'http://localhost:5000/api';
 } else if (staging) {
-  rootUrl = 'https://fhs-webapi.azurewebsites.net/api';
+  internalRrootUrl = 'https://fhs-webapi.azurewebsites.net/api';
 } else {
-  rootUrl = 'https://fhs-webapi.azurewebsites.net/api';
+  internalRrootUrl = 'https://fhs-webapi.azurewebsites.net/api';
 }
+
+const rootUrl = internalRrootUrl;
 
 export default rootUrl;
