@@ -1,4 +1,4 @@
-import Enzyme from 'enzyme';
+import * as Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 // React 16 Enzyme adapter
 Enzyme.configure({ adapter: new Adapter() });
