@@ -1,15 +1,17 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
 import ExtractCssChunks from 'extract-css-chunks-webpack-plugin';
-import HtmlWebPackPlugin from 'html-webpack-plugin';
 import HtmlTemplate from 'html-webpack-template';
 import path from 'path';
 import TsconfigPathsPlugin from 'tsconfig-paths-webpack-plugin';
 import WebpackPwaManifest from 'webpack-pwa-manifest';
 import WorkboxPlugin from 'workbox-webpack-plugin';
+import webpack from 'webpack';
+import HtmlWebPackPlugin from 'html-webpack-plugin';
 import htmlParams from '../htmlParams';
 import manifest from '../manifest';
 import paths from '../paths';
 
-const config = {
+const config: webpack.Configuration = {
   entry: {
     // client entry
     bundle: [paths.index]
